update for reservation

+++ src/main/java/com/capgemini/domain/Reservation.java
@@ -76,7 +76,7 @@ public class Reservation {
         return amountOfGuests;
     }
 
-    public void setAmountOfGuests(byte amountOfGuests) {
+    public void setAmountOfGuests(int amountOfGuests) {
         this.amountOfGuests = amountOfGuests;
     }
